Add the fci-mode to the programming modes, with the textual rule
The fill-column-indicator package displays a vertical line at the
80-character boundary. Can be toggled interactively with `SPC t f`.

The graphic rule increases the line height, so I can't see as many lines on
screen page. That's why I opted for always using the textual rule.

An issue with it: `SPC j l` shows the hint to the right of the rule for all
empty lines.