Fix the case for the first Swoop call in a session
If at the first call of Helm Swoop the thing-at-point returns an empty
string (because the cursor is not at any symbol), the original code would
throw an error saying that the symbol 'help-swoop-pattern didn't contain
any variable. That's because there was no previous Swoop input, this being
the first invocation in a session. Fix this error by returning an empty
string for Swoop to work with in that special case.