Enable visual line navigation and its left fringe indicator
Only the left indicator is enabled, because on the right we can see the
break in the fill column indicator. Plus, if a big chunk is wrapped, the
whitespace is not shown, and the visible text part sticks to the left side
of the next line, so this is where we mostly want to have the indicator.

The only reason to have the indicator in the right fringe is if a line
breaks the fill column, but is not actually wrapped due to greater window
size. Let's see if that's important enough that not having it causes
discomfort.